Allow viewing the home page without a cookie or guest user
The index page (which is really a redirect to /programs) used to be
behind a login screen. When we removed that, we created a guest user
automatically. This is problematic for web crawlers because they can't
store cookies, so it will end up redirecting to the callback to create a
guest client over and over. Additionally, this approach ends up creating
a guest user any time someone loads the site, meaning we get a lot of
empty guest accounts.

This changes the approach so that when you load / or /programs, a guest
account is not created. It is only created when you click a link to
login or start applying to a program.

Note that clicking the "End session" link while not in a session will end up creating a new
guest user. This is probably not a huge deal. When clicking "End
session" while in a session, you are dropped back to the homepage
without a new guest user being created. This could probably be
optimized.

A guest user is also created if they use program filtering. Again,
probably not a big problem, but could be optimized.