Add flag PGADMIN_CONFIG_WTF_CSRF_ENABLED = False to the dev docker container for pgadmin
This makes it so that we dont run into CSRF errors when running the tool in a web-only codespace.

Disabling CSRF should not have an impact on usages of the container in a local VS code or other environments.  Typically this would be a security issue for producion code, but since this is a dev-only tool that is only run against local dev databsaes, there should be no security concern.