Do not show gov banner when not production banner is enabled
We don't want to show the gov banner when SHOW_NOT_PRODUCTION_BANNER_ENABLED is enabled. Additionally, this fixes the banner on the J2HTML applicant UI so that toast messages render on top of it. They were previously hidden underneath it. This does not show the not production banner in the admin UI because the way the admin UI is laid out makes it more difficult, but it does hide the gov banner on the admin UI.