Allow removal of DurableJobName values
Previously, we had to keep around old durable job names in the
DurableJobName enum because if the job still existed in the database,
we'd throw an exception looking up the enum value if we didn't. This
change allows us to remove the enum value and handle the case where the
database has the old job in it, but the enum no longer has that value.