Add dependency-tree and dependency-graph commands
This adds two new bin scripts. The dependency-tree command will run dependencyTree in the SBT shell and save the output to a file called dependency-tree.txt in the base CiviForm directory. The other will run the dependencyGraphMl command, save the dependencies-compile.graphml file into the base CiviForm directory, and if it detects yEd installed on MacOS, will automatically open it in yEd.