Update generate_markdown.py to create a signed commit
Now that we require signed commits, we can't simply push a commit using a PAT like we used to. However, we can use GitHub's GraphQL API to submit a commit that is automatically signed. All commits made by the createCommitOnBranch mutation are signed by GitHub's own GPG key. This also changes the sort function so that the summary page is sorted by version, rather than textually.