[Cleanup] Remove static from buildApplicationStepDiv() used only by tests (#10650)
### Description

`buildApplicationStepDiv()` was declared as a static method in order to
call it directly from the unit test. However, that's not a good practice
since the method is no longer contained inside the class. #10640
introduced a setup method where a mock `ProgramFormBuilder` is built.
Thus, we can use it to call `buildApplicationStepDiv()` which allows us
to change the static method to a protected one.

No functionality or UI changed

### Issue(s) this completes

None