Fixing formatter to call the local docker image (#10802)
The `bin/fmt` script was running based off the docker hub tag version.
Our scripts should reference the local tag, that is the one without the
`/`.

:+1:  `civiform-formatter`
:no_good: `civiform/formatter`

This also makes it so that changes to the `bin/fmt` script will trigger
the `format` github action to re-build the formatter to take into
account any changes that may have happened.

This should now always run the formatter based on the latest changes.