Cleanup Usage of Jackson ObjectMapper Creation (#10805)
### Description

Cleaning up a mess of new `ObjectMapper` instances so I don't have to
add more to it. Best practices are to create an common instance as `new
ObjectMapper()` is a heavy-ish operation. This will also make sure we
are consistent with how we are serializing and deserializing objects.

I'm guessing we never used the default Play one because by default it
lacks support for some types.

- Replaces Play's default `ObjectMapper` configuration so that I can add
support for the Jackson modules `Jdk8Module` and `GuavaModule` which
adds support for `Optional` and Guava's Immutable collections.
- Updates classes to get the `ObjectMapper` via dependency injection
instead.

There are still a few classes that create their own. I'm leaving them
for now as they are wrapped in static methods or other harder to change
parts of the codebase. Suggest we consider a classic singleton class so
that we can get everyone using the same object, but that has more risk
and time involved that I want to handle right now.

### Checklist

#### General

Read the full guidelines for PRs
[here](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#creating-a-pull-request)

- [x] Added the correct label (see
[docs](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-appropriate-labels)
for more info): < feature | enhancement | bug | under-development |
dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more
specific round-robin
list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers)
- [ ] Added an additional reviewer from `civiform/eng-admin` as FYI (if
the primary reviewer is not an admin and this PR includes functionality
changes)
- [x] Removed the release notes section if the title is sufficient for
the release notes description, or put more details in that section.
- [ ] Created unit and/or browser tests which fail without the change
(if possible)
- [x] Performed manual testing (Chrome and Firefox if it includes
front-end changes)
- [ ] Extended the README / documentation, if necessary. For user-facing
features, consider updating [the user
docs](https://github.com/civiform/docs). For "under-the-hood" changes or
things more relevant to developers, consider updating [the dev
wiki](https://github.com/civiform/civiform/wiki).
- [ ] Ensured PII wasn't added to any new logs, unless it was guarded by
`isDevOrStaging`
- [ ] Noted in the PR description which, if any, code in this PR was
generated by AI.