[ProgramSlugUrl] Change /edit route to use program slug (#10770)
### Description

As part of #10763, we want to use the program slug instead of program id
for all program URLs on the applicant side. This PR is changes
`/programs/:programId/edit` to `/programs/:programParam/edit`

PR changes:
1. Update the route in `server/conf/routes` to receive a `programParam`
string that can be the id or slug, since play doesn't allow overloaded
controller methods and we need to support both while rolling out the
feature. It also has a `maybeProgramParam` boolean to signal the
parameter may have one (explained in later step)
2. Update edit method in
`server/app/controllers/applicant/ApplicantRoutes.java`
3. Update edit method in
`server/app/controllers/applicant/ApplicantProgramsController.java` to:
- Redirect home when feature is enabled and the program param could be a
slug but it's actually the program id (numeric). We do this instead of
making the URL invalid in case search engine indexes the URL with the
program id. We can add metrics later to see whether that happens
- Get the latest program id corresponding to the program slug when
feature is enabled and it should be a program slug. Then use it to call
`editWithApplicantId` as the current method does
- Parse the program id from the program param. Then use it to call
`editWithApplicantId` as the current method does

Since this is the first route update, we introduce new methods to
compute the latest program id corresponding to the program slug, which
is the program id of an existent draft application or otherwise the
latests active program:
1. `ApplicationRepository`: getLatestProgramId, new database query
2. `ApplicantService`: getLatestProgramId 
3. `ProgramSlugHandler`: getLatestProgramId
4. `ProgramService`: getActiveProgramId

### Checklist

#### General

- [x] Added the correct label (see
[docs](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-appropriate-labels)
for more info): < feature | enhancement | bug | under-development |
dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more
specific round-robin
list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers)
- [ ] Added an additional reviewer from `civiform/eng-admin` as FYI (if
the primary reviewer is not an admin and this PR includes functionality
changes)
- [ ] Removed the release notes section if the title is sufficient for
the release notes description, or put more details in that section.
- [x] Created unit and/or browser tests which fail without the change
(if possible)
- [x] Performed manual testing (Chrome and Firefox if it includes
front-end changes)
- [ ] Extended the README / documentation, if necessary. For user-facing
features, consider updating [the user
docs](https://github.com/civiform/docs). For "under-the-hood" changes or
things more relevant to developers, consider updating [the dev
wiki](https://github.com/civiform/civiform/wiki).
- [ ] Ensured PII wasn't added to any new logs, unless it was guarded by
`isDevOrStaging`
- [ ] Noted in the PR description which, if any, code in this PR was
generated by AI.

### Instructions for manual testing
With PROGRAM_SLUG_URLS_ENABLED off
1. Login as a Civiform admin, 
2. Create a program. Say program has slug 'programA'
3. Logout
5. Navigate to /programs/programA/edit
=> error, cannot parse a non numeric value since it's expecting a
program id

With PROGRAM_SLUG_URLS_ENABLED on
6. Navigate to /programs/programA/edit
=> redirects to an edit block page corresponding to the same program
7. Navigate to programs/:programId/edit
=> redirects back home

### Issue(s) this completes

Part of #10763