[Yes/No] Pre-populate "yes" and "no" answer options and add ability for admins to select them to be displayed to the applicant (#10832)
### Description

This PR implements two things:
1) Pre-populates "Yes" and "No" question options
2) Adds checkbox to the options which the admin can select to indicate
that the option should be displayed to applicants, or deselect to hide
from applicants.

The UI is not currently up to UX spec; that will be fixed in a follow-up
PR.

As a result of this PR, the options for future multi-select questions
(checkbox, radio, dropdown) will include a new field,
`displayInAnswerOptions`, which is set to true for each question option
when written to the database. This field is ignored for these question
types; it will only be applicable to yes/no questions.

#10776, #10778

## Release notes
As a result of this PR, the options for future multi-select questions
(checkbox, radio, dropdown) will include a new field,
`displayInAnswerOptions`, which is set to true for each question option
when written to the database. This field is ignored for these question
types; it will only be applicable to yes/no questions.

### Instructions for manual testing

1. Set YES_NO_QUESTION_ENABLED to true.
2. Log in as an admin.
3. Click "Create new question" button.
4. Click "Yes/No".
5. You should be taken to a form to create a new yes/no question, Verify
that yes/no options were pre-populated, and that if you check/uncheck
the "Include" checkbox next to each option and save, the choice is saved
when you go to edit.

Other multi option question types should be tested as well, to make sure
they weren't broken by the change.

### Issue(s) this completes

Fixes #10776