Fixing favicon (#10974)
### Description

The default favicon that was on `civiform.us` no longer exists, likely
lost when the site was redesigned. I recently noticed and have grown
tired of getting network/console errors when it fails in dev.

This will set the dev conf to not have a url set for the favicon. Now,
when unset, the favicon will default to a `data-uri` that is a white bg
with "CF" in blue text.

![image](https://github.com/user-attachments/assets/a4052d4d-cadb-49ef-81a6-265e83cc0960)


The default prod value is still set to the 404ing location on
civiform.us so that governments don't suddenly get a favicon if they
haven't set a custom one. Will take that up with Shreya to get them to
set something and eventually remove the 404ing link.

### Checklist

#### General

Read the full guidelines for PRs
[here](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#creating-a-pull-request)

- [x] Added the correct label (see
[docs](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-appropriate-labels)
for more info): < feature | enhancement | bug | under-development |
dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more
specific round-robin
list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers)
- [ ] Added an additional reviewer from `civiform/eng-admin` as FYI (if
the primary reviewer is not an admin and this PR includes functionality
changes)
- [ ] Removed the release notes section if the title is sufficient for
the release notes description, or put more details in that section.
- [ ] Created unit and/or browser tests which fail without the change
(if possible)
- [ ] Performed manual testing (Chrome and Firefox if it includes
front-end changes)
- [ ] Extended the README / documentation, if necessary. For user-facing
features, consider updating [the user
docs](https://github.com/civiform/docs). For "under-the-hood" changes or
things more relevant to developers, consider updating [the dev
wiki](https://github.com/civiform/civiform/wiki).
- [ ] Ensured PII wasn't added to any new logs, unless it was guarded by
`isDevOrStaging`
- [ ] Noted in the PR description which, if any, code in this PR was
generated by AI.