[Bug] Fix visibility check in NS end to end enumerator test (#10963)
### Description

Changes test that uses `await locator(...).isVisible()` as a test
condition (which is effectively doing nothing) to
use`expect(locator(...).toBeVisible()` so test will fail if the element
is not visible.

This surfaced an issue with the locator. `.cf-question-header` is not
existent in the page because it's not used by
[EnumeratorQuestionFragment.html](https://sourcegraph.com/github.com/civiform/civiform/-/blob/server/app/views/questiontypes/EnumeratorQuestionFragment.html).
If the goal is to verify that an enumerator entry was added, we can do
it by checking the first enumerator entry to have the expected text
value. We could have used `#enumerator-field-template-input-1` (see
[code](https://sourcegraph.com/github.com/civiform/civiform/-/blob/server/app/views/questiontypes/EnumeratorQuestionFragment.html?L78-96))
to retrieve it, but it's preferred to use `getByRole`.

### Issue(s) this completes

Part of #10955