[Bug] Fix visibility check in admin predicates test (#11096)
### Description

Changes test that uses `await locator(...).isVisible()` as a test
condition (which is effectively doing nothing) to use
`expect(locator(...).toHaveText()` to verify the items in the combobox.

This also changes getting the combobox inside `.cf-scalar-select`
instead of clicking it, which was incorrect because clicking on the
outside div was not opening the combobox. And since the `isVisible()`
was not performing a check, this part of the test was doing nothing.

### Issue(s) this completes

Part of #10955