Wire Settings Cache in the Settings Service (#11069)
### Description

Introduces an in-memory cache for admin settings to the Settings
Service. *Note:* the cache is disabled by default, and will not be
functional until a followup PR with the relevant DB evolution is merged.
#11042

## Release notes

Introduces an in-memory cache for admin settings to the Settings
Service. *Note:* the cache is disabled by default, and will not be
functional until a followup PR with the relevant DB evolution is merged.

### Checklist

#### General

Read the full guidelines for PRs
[here](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#creating-a-pull-request)

- [x] Added the correct label (see
[docs](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-appropriate-labels)
for more info): < feature | enhancement | bug | under-development |
dependencies | infrastructure | ignore-for-release | database >
- [x] Assigned to a specific person, `civiform/developers`, or a [more
specific round-robin
list](https://github.com/civiform/civiform/wiki/Technical-contribution-guide#adding-reviewers)
- [x] Added an additional reviewer from `civiform/eng-admin` as FYI (if
the primary reviewer is not an admin and this PR includes functionality
changes)
- [x] Removed the release notes section if the title is sufficient for
the release notes description, or put more details in that section.
- [ ] Created unit and/or browser tests which fail without the change
(if possible)
- [x] Performed manual testing (Chrome and Firefox if it includes
front-end changes)
- [ ] Extended the README / documentation, if necessary. For user-facing
features, consider updating [the user
docs](https://github.com/civiform/docs). For "under-the-hood" changes or
things more relevant to developers, consider updating [the dev
wiki](https://github.com/civiform/civiform/wiki).
- [ ] Ensured PII wasn't added to any new logs, unless it was guarded by
`isDevOrStaging`
- [ ] Noted in the PR description which, if any, code in this PR was
generated by AI.