Add ability to remove download options for program admins
This adds the REMOVE_DOWNLOAD_FOR_PROGRAM_ADMINS feature flag that will remove the download buttons for all program admins (but not civiform admins) when set to true. It also modifies all of the download endpoints to return unauthorized if the flag is on and the user is a program admin. This is quick fix for an urgent need by Arkansas and is intended to be temporary until we develop more fine-grained RBAC capabilities.

This also fixes a bug where the settings button was showing in the header for program admins on the application view page, because we were not setting the admin type when getting the HtmlBundle. This was not a security issue because clicking the button would just redirect and they wouldn't be able to change any settings.